Add initial android prototype
Build logic is separated into build-logic included build
The testbed project applies the playround-android plugin (the prototype plugin) from build logic
There are multiple source directories that showcase dependencies and sources are added to the correct targets
Run assemble<TargetName> to show that this all works and the correct dependencies are added to the correct targets

The prototype DSL supports hard-coded targets only
Adding targets on-demand is not supported.

There are two targets containers, targets and targets2. Each one showcases a different style where in the first,
targets are declared hard-coded by the build engineer in a static interface. In the second example, targets2, users
would theoretically be able to define targets on-demand.