Create initial unified declarative DSL plugin skeleton
Create a single build that is able to build individual Android, JVM, and KMP prototypes.
All three prototype plugins depend on a single common project to promote code sharing.

Introduces an 'Accessor' concept, which lazily applies a given plugin when the user calls a given
component 'type' using a configure closure/block. The semantics are a little odd here, since the user
needs to apply the unified plugin first, and in the Kotlin DSL must import a Kotlin Extension, but assuming
these can be fixed with changes to Gradle, we can look past this for now.

A bare minimnum implementation for each ecosystem is created. We verify that we can add dependencies to
a dependency block stub for each of Android, JVM, and KMP in both the Kotlin and Groovy DSLs.